+++ packages/backend/src/nest/storage/certificates/certificates.store.ts
@@ -147,11 +147,14 @@ export class CertificatesStore {
    * https://github.com/TryQuiet/quiet/issues/1899
    */
   protected async getCertificates() {
+    // @ts-expect-error - OrbitDB's type declaration of `load` lacks 'options'
+    await this.store.load({ fetchEntryTimeout: 15000 })
     const allCertificates = this.store
       .iterator({ limit: -1 })
       .collect()
       .map(e => e.payload.value)
 
+    logger(`All certificates: ${allCertificates.length}`)
     const validCertificates = await Promise.all(
       allCertificates.map(async certificate => {
         if (this.filteredCertificatesMapping.has(certificate)) {
@@ -159,7 +162,7 @@ export class CertificatesStore {
         }
 
         const validation = await this.validateCertificate(certificate)
-
+        logger('DuplicatedCertBug', { validation, certificate })
         if (validation) {
           const parsedCertificate = parseCertificate(certificate)
           const pubkey = keyFromCertificate(parsedCertificate)
@@ -181,7 +184,9 @@ export class CertificatesStore {
       })
     )
 
-    return validCertificates.filter(i => i != undefined)
+    const validCerts = validCertificates.filter(i => i != undefined)
+    logger(`Valid certificates: ${validCerts.length}`)
+    return validCerts
   }
 
   public async getCertificateUsername(pubkey: string) {