+++ packages/desktop/src/renderer/components/Settings/Tabs/QRCode/QRCode.stories.tsx
@@ -17,6 +17,7 @@ const invitationLink = composeInvitationShareUrl({
     },
   ],
   psk: '12345',
+  ownerOrbitDbIdentity: 'testOwnerOrbitDbIdentity',
 })
 
 const Template: ComponentStory<typeof QRCodeComponent> = args => {