+++ packages/state-manager/src/sagas/communities/communities.selectors.test.ts
@@ -41,7 +41,7 @@ describe('communitiesSelectors', () => {
 
   it('select current community', () => {
     const community = communitiesSelectors.currentCommunity(store.getState())
-    expect(community).toEqual({ ...communityAlpha, ownerCertificate: identity.userCertificate })
+    expect(community).toEqual({ ...communityAlpha })
   })
 
   it('returns registrar url without port if no port in the store', async () => {
@@ -107,12 +107,14 @@ describe('communitiesSelectors', () => {
       ),
     ]
     const psk = '12345'
+    const ownerOrbitDbIdentity = 'testOwnerOrbitDbIdentity'
     await factory.create<ReturnType<typeof communitiesActions.addNewCommunity>['payload']>('Community', {
       peerList,
+      ownerOrbitDbIdentity,
     })
     store.dispatch(communitiesActions.savePSK(psk))
     const selectorInvitationUrl = communitiesSelectors.invitationUrl(store.getState())
-    const expectedUrl = invitationShareUrl(peerList, psk)
+    const expectedUrl = invitationShareUrl(peerList, psk, ownerOrbitDbIdentity)
     expect(expectedUrl).not.toEqual('')
     expect(selectorInvitationUrl).toEqual(expectedUrl)
   })
@@ -134,44 +136,24 @@ describe('communitiesSelectors', () => {
 
   it('returns proper ownerNickname - ownerCertificate exist', async () => {
     const { store } = prepareStore()
-    expect(identity.userCertificate).not.toBeUndefined()
-    store.dispatch(
-      usersActions.responseSendCertificates({
-        certificates: [identity.userCertificate || ''],
-      })
-    )
-    const expexctedNickname = 'alice'
+    const factory = await getFactory(store)
+
     const ownerCertificate =
       'MIIDdDCCAxugAwIBAgIGAYeiqwwYMAoGCCqGSM49BAMCMA8xDTALBgNVBAMTBHRlc3QwHhcNMjMwNDIxMDcxNTMxWhcNMzAwMTMxMjMwMDAwWjBJMUcwRQYDVQQDEz5qYm1zbXR4Z2Rhd241ZTdyZ3Z5ZGZ3bGFuY3c1bnZkcmZjdXdvdmltNzJqeXY2ZTN5eDR0ZXhxZC5vbmlvbjBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABBjP55M/p8QVQGdgjtAdGwLS8uyzyIaWzvnuCvOwLs/u+FHUdb0DU2+M4TYEZjVHmqn+hSERs4XHG0/tbaaGSyGjggInMIICIzAJBgNVHRMEAjAAMAsGA1UdDwQEAwIAgDAdBgNVHSUEFjAUBggrBgEFBQcDAgYIKwYBBQUHAwEwggFHBgkqhkiG9w0BCQwEggE4BIIBNCqfocsbSvEqAdeRObiywx0KV2r7xEnqFysFuc1InEpwF3707TZGrFxww74g/ccxHCZ9zda4EHawgLoU6oKdeyec8W7qAThnWCRzJcOPINdZaTR45g28jeWAXMAtVG6eYtEQS4t7g915QaB0uYUoM3Teqp/qaMhk/9Hs4jiKlN3wL9WFYRf14XQIIVu3Fb0f3sD2/ejNPRJztJeJCwYtcFNF3fhPpH5bSPlcy6IaxhQrMXboqAfSAUlnMD4PifHFxvQYbfvTEC65Gt+FzwJ956BA5PuKsGFf+NVznyp5/YtFrl0XRRdlBcTzp2jreqhxBCdsvCpPwvM2TRv4OPk+hjMPPzBdPgvs5tytiFFyK9hXemai2TTwP1qo+VuV5SYyAyZP4rPxc/XEDHk+W3QN0vF8Ff+iMBUGCisGAQQBg4wbAgEEBxMFYWxpY2UwPQYJKwYBAgEPAwEBBDATLlFtZVN0WFY5VERXVHhoYXZUd25DZWdaYnNvMndQZ3BYQ2lzdHlCTEo2b0MyZHcwSQYDVR0RBEIwQII+amJtc210eGdkYXduNWU3cmd2eWRmd2xhbmN3NW52ZHJmY3V3b3ZpbTcyanl2NmUzeXg0dGV4cWQub25pb24wCgYIKoZIzj0EAwIDRwAwRAIga3etWmNtiMT/SUZkG0Rf5kwl3HxsGDJXsU7X5aCQAvMCIFKVBnCbTPseU5gQwamWZDG9ZoMf0X1VGzYUixWvmzuc'
+    const expectedNickname = 'alice'
 
-    const factory = await getFactory(store)
-    store.dispatch(
-      usersActions.storeUserCertificate({
-        certificate: ownerCertificate,
-      })
-    )
-    await factory.create<ReturnType<typeof communitiesActions.addOwnerCertificate>['payload']>('Community', {
+    await factory.create<ReturnType<typeof communitiesActions.updateCommunityData>['payload']>('Community', {
       ownerCertificate,
     })
+
     const ownerNickname = communitiesSelectors.ownerNickname(store.getState())
-    expect(ownerNickname).toEqual(expexctedNickname)
+
+    expect(ownerNickname).toEqual(expectedNickname)
   })
 
-  it('returns proper ownerNickname - ownerCertificate not exist - backwards compatibility', async () => {
+  it('returns proper ownerNickname - ownerCertificate not exist', async () => {
     const { store } = prepareStore()
-    expect(identity.userCertificate).not.toBeUndefined()
-    store.dispatch(
-      usersActions.responseSendCertificates({
-        certificates: [identity.userCertificate || ''],
-      })
-    )
-
-    store.dispatch(
-      usersActions.setAllCerts({
-        certificates: [identity.userCertificate || ''],
-      })
-    )
     const ownerNickname = communitiesSelectors.ownerNickname(store.getState())
-    expect(ownerNickname).toEqual(identity.nickname)
+    expect(ownerNickname).toEqual(undefined)
   })
 })